requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
bottle==0.12.18
build==0.7.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cloudify-common==6.2.0
cloudify-dsl-parser==3.0
dataclasses==0.8
extras==1.0.0
fasteners==0.16
fixtures==3.0.0
Flask==2.0.2
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==2.3.0
MarkupSafe==2.0.1
multidict==5.2.0
networkx==1.8.1
nose==1.3.7
packaging==21.3
pbr==5.8.0
pep517==0.12.0
pika==1.1.0
pkg-resources==0.0.0
pluggy==1.0.0
proxy-tools==0.1.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytz==2021.1
PyYAML==3.10
requests==2.26.0
requests-toolbelt==0.8.0
retrying==1.3.3
six==1.16.0
testtools==2.5.0
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.0
urllib3==1.26.7
wagon==0.11.0
waitress==2.0.0
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
